LayoutTests/media/modern-media-controls/time-control/time-control-expected.txt

Testing the TimeControl class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timeControl.element.localName is "div"
PASS timeControl.element.classList.contains('time-control') is true

PASS timeControl.elapsedTimeLabel is an instance of TimeLabel
PASS timeControl.scrubber is an instance of Slider
PASS timeControl.durationTimeLabel is an instance of TimeLabel
PASS timeControl.remainingTimeLabel is an instance of TimeLabel

PASS timeControl.children.length is 4
PASS timeControl.children[0] is timeControl.elapsedTimeLabel
PASS timeControl.children[1] is timeControl.scrubber
PASS timeControl.children[2] is timeControl.remainingTimeLabel
PASS timeControl.children[3] is timeControl.durationTimeLabel

timeControl.width = 500;
PASS timeControl.durationTimeLabel.x is 0
PASS getComputedStyle(timeControl.elapsedTimeLabel.element).display is not 'none'
PASS getComputedStyle(timeControl.remainingTimeLabel.element).display is 'none'
PASS getComputedStyle(timeControl.durationTimeLabel.element).display is not 'none'

timeControl.width = 100;
PASS getComputedStyle(timeControl.elapsedTimeLabel.element).display is 'none'
PASS getComputedStyle(timeControl.remainingTimeLabel.element).display is not 'none'
PASS getComputedStyle(timeControl.durationTimeLabel.element).display is 'none'

timeControl.width = 200;
PASS timeControl.elapsedTimeLabel.x is 0
PASS getComputedStyle(timeControl.elapsedTimeLabel.element).display is not 'none'
PASS getComputedStyle(timeControl.remainingTimeLabel.element).display is 'none'
PASS getComputedStyle(timeControl.durationTimeLabel.element).display is not 'none'

timeControl.durationTimeLabel.element.click();
PASS timeControl.elapsedTimeLabel.x is 0
PASS getComputedStyle(timeControl.elapsedTimeLabel.element).display is not 'none'
PASS getComputedStyle(timeControl.remainingTimeLabel.element).display is not 'none'
PASS getComputedStyle(timeControl.durationTimeLabel.element).display is 'none'

timeControl.remainingTimeLabel.element.click();
PASS timeControl.elapsedTimeLabel.x is 0
PASS getComputedStyle(timeControl.elapsedTimeLabel.element).display is not 'none'
PASS getComputedStyle(timeControl.remainingTimeLabel.element).display is 'none'
PASS getComputedStyle(timeControl.durationTimeLabel.element).display is not 'none'

PASS successfullyParsed is true

TEST COMPLETE
